+++ packages/core/src/index.spec.ts
@@ -1,4 +1,4 @@
 import { expect, test } from 'vitest';
-import { V } from './index';
+import { V } from './index.js';
 
 test('index coverage', () => expect(V).toBeDefined());
