+++ packages/core/src/index.ts
@@ -1,2 +1,4 @@
 export * from './V.js';
 export * from './validators.js';
+export * from './objectValidator.js';
+export * from './objectValidatorBuilder.js';
