+++ packages/diff/src/index.spec.ts
@@ -1,4 +1,4 @@
 import { expect, test } from 'vitest';
-import { Diff } from './index';
+import { Diff } from './index.js';
 
 test('index coverage', () => expect(Diff).toBeDefined());
