+++ packages/diff/src/index.ts
@@ -1,2 +1,2 @@
-export * from './Diff';
-export * from './VersionInfo';
+export * from './Diff.js';
+export * from './VersionInfo.js';
