+++ packages/luxon/src/index.spec.ts
@@ -1,4 +1,4 @@
 import { expect, test } from 'vitest';
-import { LuxonDateTime } from './index';
+import { LuxonDateTime } from './index.js';
 
 test('index coverage', () => expect(LuxonDateTime).toBeDefined());
