+++ packages/moment/src/index.spec.ts
@@ -1,4 +1,4 @@
 import { expect, test } from 'vitest';
-import { MomentValidator } from './index';
+import { MomentValidator } from './index.js';
 
 test('index coverage', () => expect(MomentValidator).toBeDefined());
