+++ packages/path/src/index.spec.ts
@@ -1,4 +1,4 @@
 import { expect, test } from 'vitest';
-import { Path } from './index';
+import { Path } from './index.js';
 
 test('index coverage', () => expect(Path).toBeDefined());
